Check master password length.
csgp allocate 24 bytes memory to store `cspg.pw`.
Previously the length of master password is unchecked.
Thus if user input supply a master password longer than  22 bytes
(22 + `Enter` + `NULL`), the exceeding bytes is still in stdin.
After generating and writing output,
exceeding bytes in stdin will be passed to shell,
causing the following troubles to user:

- If a user uses a master password longer than 22 bytes,
  the result (output password) of `cspg` and http://www.supergenpass.com/
  will be different.

- The exceeding bytes will be shown on command line.
  Thus if there is someone standing at the back of the user,
  they may see the last part of the master password.

- The exceeding bytes will be passed to shell:

    * If a user happens to use certain password containing valid shell command,
      then the command may be executed and be harmful,
      e.g. `22bytes+rm -rf .;more`.

    * If recording history is turned on for the shell,
      any program or people have access to shell history file may read that.

This commit checks the length of master password,
and exit the main program with a error message if length is >= 22.