detect and handle too long passwords
*csgp* operates on a small buffer (24 bytes).  a given password might be too
long. if that is the case, *csgp* reads only up to 24 bytes ... AND ignores
the rest. the rest is printed to the command line and might end up as part of
the shell history (bad).

in addition to leaking parts of the password, the output of *csgp* is
different from the original supergenpass (which is capable of processing an
"arbitrary" long master password).

to tackle both problems, this commit detects overlong passwords, drops the
rest of the password silently and exits with an error.